// lib/api/generated/model/speakRepeatDataDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { TtsDto } from './ttsDto';
import type { CheckDto } from './checkDto';

export interface SpeakRepeatDataDto {
  tts: TtsDto;
  check: CheckDto;
}